class Validation.java has moved to validation binder
